+++ src/composites.rs
@@ -1,76 +1,75 @@
 //! Module for combining hyper services
 //!
-//! Use by passing `hyper::server::NewService` instances to a `CompositeNewService`
+//! Use by passing `hyper::server::MakeService` instances to a `CompositeMakeService`
 //! together with the base path for requests that should be handled by that service.
 use futures::{future, Future};
-use hyper::server::{NewService, Service};
+use hyper::service::{MakeService, Service};
 use hyper::{Request, Response, StatusCode};
 use std::ops::{Deref, DerefMut};
 use std::{fmt, io};
 
-/// Trait for getting the path of a request. Must be implemented on the `Request`
-/// associated type for `NewService`s being combined in a `CompositeNewService`.
-pub trait GetPath {
-    /// Retrieve the path
-    fn path(&self) -> &str;
-}
-
-impl GetPath for Request {
-    fn path(&self) -> &str {
-        self.path()
-    }
-}
-
-impl<C> GetPath for (Request, C) {
-    fn path(&self) -> &str {
-        self.0.path()
-    }
-}
-
 /// Trait for generating a default "not found" response. Must be implemented on
-/// the `Response` associated type for `NewService`s being combined in a
-/// `CompositeNewService`.
-pub trait NotFound {
+/// the `Response` associated type for `MakeService`s being combined in a
+/// `CompositeMakeService`.
+pub trait NotFound<V> {
     /// Return a "not found" response
-    fn not_found() -> Self;
+    fn not_found() -> hyper::Response<V>;
 }
 
-impl NotFound for Response {
-    fn not_found() -> Self {
-        Response::new().with_status(StatusCode::NotFound)
+impl NotFound<hyper::Body> for hyper::Body {
+    fn not_found() -> hyper::Response<hyper::Body> {
+        Response::builder()
+            .status(StatusCode::NOT_FOUND)
+            .body(hyper::Body::empty())
+            .unwrap()
     }
 }
 
 type BoxedFuture<V, W> = Box<Future<Item = V, Error = W>>;
-type CompositeNewServiceVec<U, V, W> = Vec<(&'static str, Box<BoxedNewService<U, V, W>>)>;
+type CompositeMakeServiceVec<C, U, V, W> = Vec<(&'static str, Box<BoxedMakeService<C, U, V, W>>)>;
 type BoxedService<U, V, W> =
-    Box<Service<Request = U, Response = V, Error = W, Future = BoxedFuture<V, W>>>;
+    Box<Service<ReqBody = U, ResBody = V, Error = W, Future = BoxedFuture<Response<V>, W>>>;
 
-/// Trait for wrapping hyper `NewService`s to make the return type of `new_service` uniform.
-/// This is necessary in order for the `NewService`s with different `Instance` types to
+/// Trait for wrapping hyper `MakeService`s to make the return type of `make_service` uniform.
+/// This is necessary in order for the `MakeService`s with different `Instance` types to
 /// be stored in a single collection.
-pub trait BoxedNewService<U, V, W> {
+pub trait BoxedMakeService<C, U, V, W> {
     /// Create a new `Service` trait object
-    fn boxed_new_service(&self) -> Result<BoxedService<U, V, W>, io::Error>;
+    fn boxed_make_service(&mut self, context: C) -> Result<BoxedService<U, V, W>, io::Error>;
 }
 
-impl<T, U, V, W> BoxedNewService<U, V, W> for T
+impl<'a, SC, T, Rq, Rs, Er, S> BoxedMakeService<&'a SC, Rq, Rs, Er> for T
 where
-    T: NewService<Request = U, Response = V, Error = W>,
-    T::Instance: Service<Future = BoxedFuture<V, W>> + 'static,
+    S: Service<ReqBody = Rq, ResBody = Rs, Error = Er, Future = BoxedFuture<Response<Rs>, Er>>
+        + 'static,
+    T: MakeService<
+        &'a SC,
+        ReqBody = Rq,
+        ResBody = Rs,
+        Error = Er,
+        Future = futures::future::FutureResult<S, io::Error>,
+        Service = S,
+        MakeError = io::Error,
+    >,
+    Rq: hyper::body::Payload,
+    Rs: hyper::body::Payload,
+    Er: std::error::Error + Send + Sync + 'static,
 {
-    /// Call the `new_service` method of the wrapped `NewService` and `Box` the result
-    fn boxed_new_service(&self) -> Result<BoxedService<U, V, W>, io::Error> {
-        let service = self.new_service()?;
+    /// Call the `make_service` method of the wrapped `MakeService` and `Box` the result
+    fn boxed_make_service(
+        &mut self,
+        context: &'a SC,
+    ) -> Result<BoxedService<Rq, Rs, Er>, io::Error> {
+        let service = self.make_service(context).wait()?;
         Ok(Box::new(service))
     }
 }
 
 /// Wraps a vector of pairs, each consisting of a base path as a `&'static str`
-/// and a `NewService` instance. Implements `Deref<Vec>` and `DerefMut<Vec>` so
+/// and a `MakeService` instance. Implements `Deref<Vec>` and `DerefMut<Vec>` so
 /// these can be manipulated using standard `Vec` methods.
 ///
-/// The `Service` returned by calling `new_service()` will pass an incoming
+/// The `Service` returned by calling `make_service()` will pass an incoming
 /// request to the first `Service` in the list for which the associated
 /// base path is a prefix of the request path.
 ///
@@ -78,20 +77,19 @@ where
 /// =============
 ///
 /// ```ignore
-/// let my_new_service1 = NewService1::new();
-/// let my_new_service2 = NewService2::new();
+/// let my_make_service1 = MakeService1::new();
+/// let my_make_service2 = MakeService2::new();
 ///
-/// let mut composite_new_service = CompositeNewService::new();
-/// composite_new_service.push(("/base/path/1", my_new_service1));
-/// composite_new_service.push(("/base/path/2", my_new_service2));
+/// let mut composite_make_service = CompositeMakeService::new();
+/// composite_make_service.push(("/base/path/1", my_make_service1));
+/// composite_make_service.push(("/base/path/2", my_make_service2));
 ///
-/// // use as you would any `NewService` instance
+/// // use as you would any `MakeService` instance
 /// ```
 #[derive(Default)]
-pub struct CompositeNewService<U, V, W>(CompositeNewServiceVec<U, V, W>)
+pub struct CompositeMakeService<C, U, V, W>(CompositeMakeServiceVec<C, U, V, W>)
 where
-    U: GetPath,
-    V: NotFound + 'static,
+    V: NotFound<V> + 'static,
     W: 'static;
 
 // Workaround for https://github.com/rust-lang-nursery/rust-clippy/issues/2226
@@ -103,64 +101,75 @@ where
         clippy::new_without_default_derive
     )
 )]
-impl<U: GetPath, V: NotFound, W> CompositeNewService<U, V, W> {
-    /// create an empty `CompositeNewService`
+impl<C, U, V: NotFound<V>, W> CompositeMakeService<C, U, V, W> {
+    /// create an empty `CompositeMakeService`
     pub fn new() -> Self {
-        CompositeNewService(Vec::new())
+        CompositeMakeService(Vec::new())
     }
 }
 
-impl<U, V, W> NewService for CompositeNewService<U, V, W>
+impl<'a, C, U, V, W> MakeService<&'a C> for CompositeMakeService<&'a C, U, V, W>
 where
-    U: GetPath,
-    V: NotFound + 'static,
-    W: 'static,
+    U: hyper::body::Payload,
+    V: NotFound<V> + 'static + hyper::body::Payload,
+    W: std::error::Error + Send + Sync + 'static,
 {
-    type Request = U;
-    type Response = V;
+    type ReqBody = U;
+    type ResBody = V;
     type Error = W;
-    type Instance = CompositeService<U, V, W>;
+    type Service = CompositeService<U, V, W>;
+    type MakeError = io::Error;
+    type Future = futures::future::FutureResult<Self::Service, io::Error>;
 
-    fn new_service(&self) -> Result<Self::Instance, io::Error> {
+    fn make_service(
+        &mut self,
+        _service_ctx: &'a C,
+    ) -> futures::future::FutureResult<Self::Service, io::Error> {
         let mut vec = Vec::new();
 
-        for &(base_path, ref new_service) in &self.0 {
-            vec.push((base_path, new_service.boxed_new_service()?))
+        for &mut (base_path, ref mut make_service) in &mut self.0 {
+            vec.push((
+                base_path,
+                make_service
+                    .boxed_make_service(_service_ctx)
+                    .expect("Error"),
+            ))
         }
 
-        Ok(CompositeService(vec))
+        future::FutureResult::from(Ok(CompositeService(vec)))
     }
 }
 
-impl<U, V, W> fmt::Debug for CompositeNewService<U, V, W>
+impl<C, U, V, W> fmt::Debug for CompositeMakeService<C, U, V, W>
 where
-    U: GetPath,
-    V: NotFound + 'static,
+    V: NotFound<V> + 'static,
     W: 'static,
 {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         // Get vector of base paths
         let str_vec: Vec<&'static str> = self.0.iter().map(|&(base_path, _)| base_path).collect();
-        write!(f, "CompositeNewService accepting base paths: {:?}", str_vec,)
+        write!(
+            f,
+            "CompositeMakeService accepting base paths: {:?}",
+            str_vec,
+        )
     }
 }
 
-impl<U, V, W> Deref for CompositeNewService<U, V, W>
+impl<C, U, V, W> Deref for CompositeMakeService<C, U, V, W>
 where
-    U: GetPath,
-    V: NotFound + 'static,
+    V: NotFound<V> + 'static,
     W: 'static,
 {
-    type Target = CompositeNewServiceVec<U, V, W>;
+    type Target = CompositeMakeServiceVec<C, U, V, W>;
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
 
-impl<U, V, W> DerefMut for CompositeNewService<U, V, W>
+impl<C, U, V, W> DerefMut for CompositeMakeService<C, U, V, W>
 where
-    U: GetPath,
-    V: NotFound + 'static,
+    V: NotFound<V> + 'static,
     W: 'static,
 {
     fn deref_mut(&mut self) -> &mut Self::Target {
@@ -172,26 +181,25 @@ where
 /// and a `Service` instance.
 pub struct CompositeService<U, V, W>(Vec<(&'static str, BoxedService<U, V, W>)>)
 where
-    U: GetPath,
-    V: NotFound + 'static,
+    V: NotFound<V> + 'static,
     W: 'static;
 
 impl<U, V, W> Service for CompositeService<U, V, W>
 where
-    U: GetPath,
-    V: NotFound + 'static,
-    W: 'static,
+    U: hyper::body::Payload,
+    V: NotFound<V> + 'static + hyper::body::Payload,
+    W: 'static + std::error::Error + Send + Sync,
 {
-    type Request = U;
-    type Response = V;
+    type ReqBody = U;
+    type ResBody = V;
     type Error = W;
-    type Future = Box<Future<Item = V, Error = W>>;
+    type Future = Box<Future<Item = Response<V>, Error = W>>;
 
-    fn call(&self, req: Self::Request) -> Self::Future {
+    fn call(&mut self, req: Request<Self::ReqBody>) -> Self::Future {
         let mut result = None;
 
-        for &(base_path, ref service) in &self.0 {
-            if req.path().starts_with(base_path) {
+        for &mut (base_path, ref mut service) in &mut self.0 {
+            if req.uri().path().starts_with(base_path) {
                 result = Some(service.call(req));
                 break;
             }
@@ -203,8 +211,7 @@ where
 
 impl<U, V, W> fmt::Debug for CompositeService<U, V, W>
 where
-    U: GetPath,
-    V: NotFound + 'static,
+    V: NotFound<V> + 'static,
     W: 'static,
 {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
@@ -216,8 +223,7 @@ where
 
 impl<U, V, W> Deref for CompositeService<U, V, W>
 where
-    U: GetPath,
-    V: NotFound + 'static,
+    V: NotFound<V> + 'static,
     W: 'static,
 {
     type Target = Vec<(&'static str, BoxedService<U, V, W>)>;
@@ -228,8 +234,7 @@ where
 
 impl<U, V, W> DerefMut for CompositeService<U, V, W>
 where
-    U: GetPath,
-    V: NotFound + 'static,
+    V: NotFound<V> + 'static,
     W: 'static,
 {
     fn deref_mut(&mut self) -> &mut Self::Target {